Remove redundant slavetemplate tests
